% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC}
\alias{AUC}
\title{Caculate AUC}
\usage{
AUC(T.Beta, Beta)
}
\arguments{
\item{T.Beta}{True label}

\item{Beta}{estimated coefficient}
}
\value{
X generated data
}
\description{
Caculate AUC
}
\examples{
 AUC(T.Beta = c(1,0,1), Beta <- c(0.4,0.7,0.1))


}
