% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto.R
\name{Auto}
\alias{Auto}
\title{Generate a normal distributed sample X with dimension n *p with mean 0
and an autocorrelation covariance matrix where a_{ij} = rho^{i-j}}
\usage{
Auto(rho, P, N)
}
\arguments{
\item{rho}{auto correlation structure}

\item{P}{dimension of the data}

\item{N}{sample size of the data}
}
\value{
X generated data
}
\description{
Generate a normal distributed sample X with dimension n *p with mean 0
and an autocorrelation covariance matrix where a_{ij} = rho^{i-j}
}
\examples{
 Auto(rho = 0.2, P = 10, N = 300)

}
