% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Betacoef.R
\name{beta.coef}
\alias{beta.coef}
\title{Generate a sparse vector whose non-zero elements following
U(0,1) distribution.}
\usage{
beta.coef(K, P)
}
\arguments{
\item{K}{number of non-zero elements}

\item{P}{length of the vector}
}
\value{
A sparse vector
}
\description{
Generate a sparse vector whose non-zero elements following
U(0,1) distribution.
}
\examples{
 beta.coef(K = 1, P = 10)

}
