% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block.R
\name{Block}
\alias{Block}
\title{Generate a normal distributed sample X with dimension n *p with mean 0
and a blockdiagonal covariance matrix}
\usage{
Block(rho, P, N, bandwidth)
}
\arguments{
\item{rho}{off-diagnoal elements value}

\item{P}{dimension of the data}

\item{N}{sample size of the data}

\item{bandwidth}{blok size}
}
\value{
X generated data
}
\description{
Generate a normal distributed sample X with dimension n *p with mean 0
and a blockdiagonal covariance matrix
}
\examples{
 Block(rho = 0.2, P = 10, N = 300, bandwidth = 5)

}
