% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmlasso.R
\name{glmreg}
\alias{glmreg}
\title{fit a model with given X and y with
precision lasso method}
\usage{
glmreg(X, y, logistic = TRUE, method = c("lasso", "ridge",
  "elasticnet"))
}
\arguments{
\item{X}{auto correlation structure}

\item{y}{dimension of the data}

\item{logistic}{indicator for doing regression on binary response}

\item{method}{choose the method: lasso, ridge or elasticnet}
}
\value{
coefficients generated
}
\description{
fit a model with given X and y with
precision lasso method
}
\examples{
 X = matrix(rnorm(60*50), ncol = 60, nrow = 50)
 y = rnorm(50)
 lmcoef = glmreg(X, y, logistic = FALSE, method = "lasso")
}
