% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleSimulation.R
\name{multipleSimulation}
\alias{multipleSimulation}
\title{Multiple Simulation}
\usage{
multipleSimulation(type.vec, rho.vec, n.vec, K.vec, mccores = 1)
}
\arguments{
\item{type.vec}{covariance parameter vector}

\item{rho.vec}{correlation parameter vector}

\item{n.vec}{parameter vector of sample size}

\item{K.vec}{parameter vector of sparsity}

\item{mccores}{specify the cores number for parallel computing. Default is 1}
}
\value{
calculated AUC tables for different combination
}
\description{
Multiple Simulation
}
\examples{
 multipleSimulation(type.vec = c("Auto"),
  rho.vec = c(0.2),
   n.vec =c(100),
    K.vec = c(20))

}
