% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pslasso.R
\name{PSlasso}
\alias{PSlasso}
\title{fit a model with given X and y with
precision lasso method}
\usage{
PSlasso(X, y,
  path = "/Users/w/Desktop/UPenn Course/2019 spring/computing/pslassoR/R/",
  logistic = TRUE)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{path}{abslute path where the python file is stored}

\item{logistic}{indicator for doing regression on binary response}
}
\value{
coefficients generated
}
\description{
fit a model with given X and y with
precision lasso method
}
\examples{
 X = matrix(rnorm(60*50), ncol = 60, nrow = 50)
 y = rnorm(50)
 lmcoef = PSlasso(X, y, logistic = TRUE)

}
