% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Simulation}
\alias{Simulation}
\title{Calculate the AUC for five variable selection method}
\usage{
Simulation(i, n, K, rho, type, path)
}
\arguments{
\item{i}{random seed}

\item{n}{dimension of the data}

\item{K}{Sparsity parameter}

\item{rho}{parameter of covariance matrix}

\item{type}{type of covariance matrix}
}
\value{
calculated AUC for five methods
}
\description{
Calculate the AUC for five variable selection method
}
\examples{
 Simulation(i = 1, n = 100, K =10, rho = 0.2,
 type = "block")

}
