% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraceLasso.R
\name{TraceLasso}
\alias{TraceLasso}
\title{fit a model with given X and y
 with trace lasso method}
\usage{
TraceLasso(X, y,
  path = "/Users/w/Desktop/UPenn Course/2019 spring/computing/pslassoR/R/",
  logistic = TRUE)
}
\arguments{
\item{X}{auto correlation structure}

\item{y}{dimension of the data}

\item{path}{abslute path where the python file is stored}

\item{logistic}{indicator for using logistic or not}
}
\value{
coefficients generated
}
\description{
fit a model with given X and y
 with trace lasso method
}
\examples{
 X = matrix(rnorm(60*50), ncol = 60, nrow = 50)
 y = rnorm(50)
 lmcoef = TraceLasso(X, y, logistic = TRUE)

}
